+++ lst-optimizer/packages/lst-optimizer-std/src/types/pool_asset.rs
@@ -1,12 +1,12 @@
 #[derive(Debug, Clone)]
 pub struct PoolAsset {
     pub mint: String,
-    pub lamports: i128,
+    pub lamports: u64,
     pub reserves: u64,
 }
 
 impl PoolAsset {
-    pub fn new(mint: &str, lamports: i128, reserves: u64) -> Self {
+    pub fn new(mint: &str, lamports: u64, reserves: u64) -> Self {
         Self { mint: mint.to_string(), lamports, reserves }
     }
 }
